+++ electron-builder.config.js
@@ -60,8 +60,7 @@ const config = {
     '**/node_modules/sharp/**',
     '**/node_modules/semver/**',
     '**/node_modules/detect-libc/**',
-    '**/node_modules/@img/**',
-    '**/node_modules/ffmpeg-static/**'
+    '**/node_modules/@img/**'
   ],
   compression: 'maximum',
   nsis: {